chore: do not run external resource-dependent publishes in forks
Currently, when forking the repository and trying to publish a new version via CI (a github release) the CI run fails due to missing credentials.

This disables publishes for Cargo, NPM and into the dprint/plugins repo when running the CI step on a fork.
The only publish that is untouched is the github release, as each fork would have that available by default.

The other option to make this happen is to embed the existence of the secrets (`CI_REPO_PAT`, `CRATES_TOKEN`, `NPM_TOKEN`) into the `if` condition. I think it is less obvious, but let me know if you prefer that.