Cast &last_node to &dyn Spanned
This change fixes a todo in the parser, replacing an `if` statement that
has an implicit cast to a one-liner and an explicit cast.

The function `parse_comments_as_statements()` expects a `last_node:
Option<&dyn Spanned>`. But `last_node.as_ref()` will map an
`Option<Span>` to an `Option<&Span>`. IIUC, this results in a compiler
error as `&Span` and `&dyn Span` have different memory layouts (`&Span`
is a struct that is statically dispatched, and `&dyn Span` is a trait
object that is dynamically dispatched) and so are incompatible types.

With the `if` statement before, the compiler does this cast for you
implicitly.