Replace CharIterator struct with IteratorCharExt trait
This change replaces the `CharIterator` struct with the `IteratorCharExt` trait that is implemented for `Peekable<Chars>`. This has the benefits that

- we don't have to wrap `Chars` as `CharIterator`, and subsequently implement iterator peeking ourselves. Instead we can rely on the `Peekable` struct provided in `std::iter`; and
- we are able to extend the functionality of `Peekable<Chars>` (and any other types that implement `IteratorCharExt`).

I hope this change accurately actions the todo. Happy to work through additional improvements or ways we can make this more flexible if there are areas that would benefit from its reuse.