fix: Improve contextual spacing for if statements when ifStatement.spaceAround: true
When ifStatement.spaceAround is true, make sure the spaces around parens are right based on the context of where the paren lies.

The change is to only add the space if the open paren is not at the end of a line. We do that on the opening paren by adding a Signal::SpaceOrNewLine which allows the system to add the space if it needs to for purposes of word wrap, otherwise ommitting it. For the close paren, we add the space if we're not the first token on the line.

The impetus for this was the case where an opening paren for an if statement control block was the last character on a line, or a closing paren was the first.
For example: (where lineWidth: 40)
```
if (
    somereallylongidentifier
        <= anotherreallylongidentifier
) {
}
```

The above case was getting formatted with an extra space at the end of the if statement line and an extra space at the start of the close paren line.